Add support for R_max/2 fits
Add option "R_max/2" in addition to "richards" and "gauss_cdf" as
"fitting function". Choosing this keyword as "fit function" will
simply use linear interpolation between the two datapoints closest to
where the superconducting transition happens (by definition at R_max/2,
so where the resistance is half the normal conducting resistance).
This is especially intended for use with sweeps that have a high
resolution in temperature such that a fit is unnecessary or even
yielding worse results.